+++ fuc/api/pycov.py
@@ -858,26 +858,36 @@ class CovFrame:
         >>> import numpy as np
         >>> from fuc import pycov
         >>> data = {
-        ...     'Chromosome': ['chr1'] * 1000,
-        ...     'Position': np.arange(1000, 2000),
-        ...     'A': pycov.simulate(loc=35, scale=5),
-        ...     'B': pycov.simulate(loc=25, scale=7),
+        ...     'Chromosome': ['chr1'] * 3 + ['2'] * 3,
+        ...     'Position': np.arange(1, 7),
+        ...     'A': pycov.simulate(loc=35, scale=5, size=6),
+        ...     'B': pycov.simulate(loc=25, scale=7, size=6),
         ... }
         >>> cf = pycov.CovFrame.from_dict(data)
-        >>> cf.df.head()
+        >>> cf.df
           Chromosome  Position   A   B
-        0       chr1      1000  28  17
-        1       chr1      1001  35  28
-        2       chr1      1002  38  12
-        3       chr1      1003  33  20
-        4       chr1      1004  33  31
-        >>> cf.chr_prefix().df.head()
+        0       chr1         1  35  25
+        1       chr1         2  23  14
+        2       chr1         3  32  23
+        3          2         4  38  25
+        4          2         5  33   8
+        5          2         6  21  22
+        >>> cf.chr_prefix(mode='remove').df
           Chromosome  Position   A   B
-        0          1      1000  28  17
-        1          1      1001  35  28
-        2          1      1002  38  12
-        3          1      1003  33  20
-        4          1      1004  33  31
+        0          1         1  35  25
+        1          1         2  23  14
+        2          1         3  32  23
+        3          2         4  38  25
+        4          2         5  33   8
+        5          2         6  21  22
+        >>> cf.chr_prefix(mode='add').df
+          Chromosome  Position   A   B
+        0       chr1         1  35  25
+        1       chr1         2  23  14
+        2       chr1         3  32  23
+        3       chr2         4  38  25
+        4       chr2         5  33   8
+        5       chr2         6  21  22
         """
         if mode == 'remove':
             def one_row(r):
@@ -885,7 +895,8 @@ class CovFrame:
                 return r
         elif mode == 'add':
             def one_row(r):
-                r.Chromosome = 'chr' + r.Chromosome
+                if 'chr' not in r.Chromosome:
+                    r.Chromosome = 'chr' + r.Chromosome
                 return r
         else:
             raise ValueError(f'Incorrect mode: {mode}')
@@ -957,3 +968,61 @@ class CovFrame:
             samples = [x for x in self.samples if x not in samples]
         cols = self.df.columns[:2].to_list() + samples
         return self.__class__(self.df[cols])
+
+    def rename(self, names, indicies=None):
+        """
+        Rename the samples.
+
+        Parameters
+        ----------
+        names : dict or list
+            Dict of old names to new names or list of new names.
+        indicies : list or tuple, optional
+            List of 0-based sample indicies. Alternatively, a tuple
+            (int, int) can be used to specify an index range.
+
+        Returns
+        -------
+        CovFrame
+            Updated CovFrame.
+
+        Examples
+        --------
+
+        >>> import numpy as np
+        >>> from fuc import pycov
+        >>> data = {
+        ...     'Chromosome': ['chr1'] * 2,
+        ...     'Position': np.arange(1, 3),
+        ...     'A': pycov.simulate(loc=35, scale=5, size=2),
+        ...     'B': pycov.simulate(loc=25, scale=7, size=2),
+        ...     'C': pycov.simulate(loc=25, scale=7, size=2),
+        ...     'D': pycov.simulate(loc=25, scale=7, size=2),
+        ... }
+        >>> cf = pycov.CovFrame.from_dict(data)
+        >>> cf.df
+          Chromosome  Position   A   B   C   D
+        0       chr1         1  31  19  28  15
+        1       chr1         2  35  24  22  17
+        >>> cf.rename(['1', '2', '3', '4']).df
+          Chromosome  Position   1   2   3   4
+        0       chr1         1  31  19  28  15
+        1       chr1         2  35  24  22  17
+        >>> cf.rename({'B': '2', 'C': '3'}).df
+          Chromosome  Position   A   2   3   D
+        0       chr1         1  31  19  28  15
+        1       chr1         2  35  24  22  17
+        >>> cf.rename(['2', '4'], indicies=[1, 3]).df
+          Chromosome  Position   A   2   C   4
+        0       chr1         1  31  19  28  15
+        1       chr1         2  35  24  22  17
+        >>> cf.rename(['2', '3'], indicies=(1, 3)).df
+          Chromosome  Position   A   2   3   D
+        0       chr1         1  31  19  28  15
+        1       chr1         2  35  24  22  17
+        """
+        samples = common.rename(self.samples, names, indicies=indicies)
+        columns = self.df.columns[:2].to_list() + samples
+        cf = self.copy()
+        cf.df.columns = columns
+        return cf